requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argcomplete==3.2.2
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
distro==1.9.0
fastdtw==0.3.4
filelock==3.13.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.1.0
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-speech==2.24.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
hydra-core==1.3.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mecab-python3==1.0.8
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.2.1
nnmnkwii==0.1.2
numba==0.59.0
numpy==1.26.4
omegaconf==2.3.0
openai==0.28.0
packaging==23.2
parallel-wavegan==0.6.1
pika==1.3.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyopenjtalk==0.3.3
pyparsing==3.1.1
PySocks==1.7.1
pysptk==0.2.2
python-dateutil==2.8.2
pyworld==0.3.4
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tomlkit==0.12.3
torch==2.2.0
tqdm==4.66.1
ttslearn==0.2.2
typing_extensions==4.9.0
unidic-lite==1.0.8
urllib3==2.2.0
xmltodict==0.13.0
yarl==1.9.4
yq==3.2.3
